Add external folders to resolve import conflicts
