+++ cardano-server.cabal
@@ -31,29 +31,29 @@ common lang
 library
     import: lang
     exposed-modules: 
-      Client.Class
-      Client.Client
-      Client.Default
-      Client.Opts
-      Server.Class
-      Server.Config
-      Server.Endpoints.AddSignature
-      Server.Endpoints.Funds 
-      Server.Endpoints.Ping
-      Server.Endpoints.Servant
-      Server.Endpoints.Tx.Class
-      Server.Endpoints.Tx.Internal
-      Server.Endpoints.Tx.New
-      Server.Endpoints.Tx.Submit      
-      Server.Error
-      Server.Internal
-      Server.Main
-      Server.Tx
-      TestingServer.Main
-      TestingServer.OnChain
-      TestingServer.OffChain
-      Utils.Logger
-      Utils.Wait
+      Cardano.Server.Client.Class
+      Cardano.Server.Client.Client
+      Cardano.Server.Client.Default
+      Cardano.Server.Client.Opts
+      Cardano.Server.Class
+      Cardano.Server.Config
+      Cardano.Server.Endpoints.AddSignature
+      Cardano.Server.Endpoints.Funds 
+      Cardano.Server.Endpoints.Ping
+      Cardano.Server.Endpoints.Servant
+      Cardano.Server.Endpoints.Tx.Class
+      Cardano.Server.Endpoints.Tx.Internal
+      Cardano.Server.Endpoints.Tx.New
+      Cardano.Server.Endpoints.Tx.Submit      
+      Cardano.Server.Error
+      Cardano.Server.Internal
+      Cardano.Server.Main
+      Cardano.Server.TestingServer.Main
+      Cardano.Server.TestingServer.OnChain
+      Cardano.Server.TestingServer.OffChain
+      Cardano.Server.Tx
+      Cardano.Server.Utils.Logger
+      Cardano.Server.Utils.Wait
     other-modules:
     default-extensions:
       PartialTypeSignatures
