+++ src/Cardano/Server/Client/Default.hs
@@ -5,17 +5,18 @@
 {-# LANGUAGE ScopedTypeVariables        #-}
 {-# LANGUAGE TypeApplications           #-}
 
-module Client.Default where
+module Cardano.Server.Client.Default where
 
-import           Client.Client             (mkRequest)
-import           Control.Monad             (void)
-import           Data.Aeson                (FromJSON)
-import qualified Data.Text                 as T
-import           Network.HTTP.Client       (defaultManagerSettings, newManager)
-import           Server.Internal           (runAppM, HasServer(..))
-import           Server.Config             (Config(..), loadConfig, decodeOrErrorFromFile)       
-import           Options.Applicative       ((<**>), auto, fullDesc, help, info, long, option, short, value, execParser, helper)
-import           Utils.ChainIndex          (MapUTXO)
+import           Cardano.Server.Client.Client (mkRequest)
+import           Cardano.Server.Config        (Config(..), loadConfig, decodeOrErrorFromFile)       
+import           Cardano.Server.Internal      (runAppM, HasServer(..))
+import           Control.Monad                (void)
+import           Data.Aeson                   (FromJSON)
+import qualified Data.Text                    as T
+import           Network.HTTP.Client          (defaultManagerSettings, newManager)
+import           Options.Applicative          ((<**>), auto, fullDesc, help, info, long, option, short, value,
+                                               execParser, helper)
+import           Utils.ChainIndex             (MapUTXO)
 
 -- Running a client that only needs fromJSON instance of server input
 -- instead of defining a full HasClient class.
