+++ src/Cardano/Server/Error.hs
@@ -5,15 +5,15 @@
 {-# LANGUAGE RankNTypes        #-}
 {-# LANGUAGE TypeApplications  #-}
 
-module Server.Error where
+module Cardano.Server.Error where
 
-import           Data.Text                (Text)
-import           Control.Monad.Catch      (Exception(..), MonadThrow(..), handle)
-import           IO.ChainIndex            (pattern ChainIndexConnectionError)
-import           IO.Wallet                (pattern WalletApiConnectionError)
-import           Server.Internal          (NetworkM)
-import           Servant                  (WithStatus, IsMember, Union)
-import           Server.Endpoints.Servant (respondWithStatus)
+import           Cardano.Server.Endpoints.Servant (respondWithStatus)
+import           Cardano.Server.Internal          (NetworkM)
+import           Control.Monad.Catch              (Exception(..), MonadThrow(..), handle)
+import           Data.Text                        (Text)
+import           IO.ChainIndex                    (pattern ChainIndexConnectionError)
+import           IO.Wallet                        (pattern WalletApiConnectionError)
+import           Servant                          (WithStatus, IsMember, Union)
 
 handleUnavailableEndpoints :: forall s res. (IsMember (WithStatus 503 Text) res) => 
     NetworkM s (Union res) -> NetworkM s (Union res)
