+++ src/Cardano/Server/Internal.hs
@@ -6,8 +6,8 @@
 {-# LANGUAGE ScopedTypeVariables        #-}
 {-# LANGUAGE TypeApplications           #-}
 
-module Server.Internal 
-    ( module Server.Class
+module Cardano.Server.Internal 
+    ( module Cardano.Server.Class
     , NetworkM(..)
     , AppM(..)
     , runAppM
@@ -15,18 +15,18 @@ module Server.Internal
     , loadEnv
     ) where
 
-import           Control.Monad.Catch    (MonadThrow, MonadCatch)
-import           Control.Monad.IO.Class (MonadIO)
-import           Control.Monad.Reader   (ReaderT(ReaderT, runReaderT), MonadReader, asks)
-import           Data.Default           (def)
-import           Data.IORef             (newIORef)
-import           Data.Sequence          (empty)
-import           IO.Wallet              (HasWallet(..))
-import           Ledger                 (Params(..), pParamsFromProtocolParams)
-import           Servant                (Handler)
-import           Server.Class           (Env(..), Queue, QueueRef, QueueElem, HasServer(..))
-import           Server.Config          (Config(..), configFile, decodeOrErrorFromFile)
-import           Utils.Logger           (HasLogger(..))
+import           Cardano.Server.Class        (Env(..), Queue, QueueRef, QueueElem, HasServer(..))
+import           Cardano.Server.Config       (Config(..), configFile, decodeOrErrorFromFile)
+import           Cardano.Server.Utils.Logger (HasLogger(..))
+import           Control.Monad.Catch         (MonadThrow, MonadCatch)
+import           Control.Monad.IO.Class      (MonadIO)
+import           Control.Monad.Reader        (ReaderT(ReaderT, runReaderT), MonadReader, asks)
+import           Data.Default                (def)
+import           Data.IORef                  (newIORef)
+import           Data.Sequence               (empty)
+import           IO.Wallet                   (HasWallet(..))
+import           Ledger                      (Params(..), pParamsFromProtocolParams)
+import           Servant                     (Handler)
 
 newtype NetworkM s a = NetworkM { unNetworkM :: ReaderT (Env s) Handler a }
     deriving newtype
