+++ test/Tests/DefaultClient.hs
@@ -5,15 +5,15 @@
 
 module Tests.DefaultClient where
 
-import           Client.Class           (HasClient(..))
-import           Client.Default         (defaultClient)
-import           Control.Exception      (handle, SomeException)
-import           Control.Monad.IO.Class (MonadIO(..))
-import           Data.Aeson             (encode, FromJSON)
-import qualified Data.ByteString.Lazy   as LBS
-import           Server.Class           (HasServer(InputOf))
-import           Server.Internal        (runAppM)
-import           System.Directory       (removeFile)
+import           Cardano.Server.Class          (HasServer(InputOf))
+import           Cardano.Server.Client.Class   (HasClient(..))
+import           Cardano.Server.Client.Default (defaultClient)
+import           Cardano.Server.Internal       (runAppM)
+import           Control.Exception             (handle, SomeException)
+import           Control.Monad.IO.Class        (MonadIO(..))
+import           Data.Aeson                    (encode, FromJSON)
+import qualified Data.ByteString.Lazy          as LBS
+import           System.Directory              (removeFile)
 
 testDefaultClient :: forall s. (HasClient s, FromJSON (InputOf s)) => IO ()
 testDefaultClient = runAppM @s $ do
