+++ test/Tests/Reference.hs
@@ -5,19 +5,19 @@
 
 module Tests.Reference where
 
-import           Control.Monad           (void)
-import qualified Data.Map                as Map
-import           IO.Wallet               (getWalletAddr)
-import qualified Ledger.Ada              as Ada
-import           Ledger.Typed.Scripts    (Any)
-import           Constraints.OffChain    (postMintingPolicyTx, referenceMintingPolicyTx)
-import           Server.Internal         (runAppM)
-import           Server.Tx               (mkTx)
-import           TestingServer.Main      (TestingServer)
-import           TestingServer.OffChain  (testToken)
-import           TestingServer.OnChain   (testPolicyV, testPolicy)
-import qualified PlutusTx.Prelude        as Plutus
-import           Utils.Logger            (HasLogger(..))
+import           Cardano.Server.Internal                (runAppM)
+import           Cardano.Server.TestingServer.Main      (TestingServer)
+import           Cardano.Server.TestingServer.OffChain  (testToken)
+import           Cardano.Server.TestingServer.OnChain   (testPolicyV, testPolicy)
+import           Cardano.Server.Tx                      (mkTx)
+import           Cardano.Server.Utils.Logger            (HasLogger(..))
+import           Constraints.OffChain                   (postMintingPolicyTx, referenceMintingPolicyTx)
+import           Control.Monad                          (void)
+import qualified Data.Map                               as Map
+import           IO.Wallet                              (getWalletAddr)
+import qualified Ledger.Ada                             as Ada
+import           Ledger.Typed.Scripts                   (Any)
+import qualified PlutusTx.Prelude                       as Plutus
 
 postReferenceScript :: IO ()
 postReferenceScript = void $ runAppM @TestingServer $ do
