+++ test/Tests/TestingServer.hs
@@ -4,19 +4,19 @@
 
 module Tests.TestingServer where
 
-import           Control.Monad.IO.Class     (MonadIO(..))
-import qualified Data.Map                   as Map
-import           IO.ChainIndex              (getUtxosAt)
-import           IO.Wallet                  (getWalletAddr)
-import           PlutusTx.Builtins.Class    (stringToBuiltinByteString)
-import           Server.Internal            (HasServer(..), runAppM)
-import           Server.Endpoints.Tx.Submit (processTokens)
-import           Server.Tx                  (mkWalletTxOutRefs) 
-import           Tests.DefaultClient        (testDefaultClient)
-import           Tests.Internal             (testFunds, testFundsAll)
-import           TestingServer.Main         (TestingServer)
-import           Utils.ChainIndex           (filterCleanUtxos)
-import           Utils.Logger               (HasLogger(..), logSmth, (.<))
+import           Cardano.Server.Endpoints.Tx.Submit (processTokens)
+import           Cardano.Server.Internal            (HasServer(..), runAppM)
+import           Cardano.Server.TestingServer.Main  (TestingServer)
+import           Cardano.Server.Tx                  (mkWalletTxOutRefs) 
+import           Cardano.Server.Utils.Logger        (HasLogger(..), logSmth, (.<))
+import           Control.Monad.IO.Class             (MonadIO(..))
+import qualified Data.Map                           as Map
+import           IO.ChainIndex                      (getUtxosAt)
+import           IO.Wallet                          (getWalletAddr)
+import           PlutusTx.Builtins.Class            (stringToBuiltinByteString)
+import           Tests.DefaultClient                (testDefaultClient)
+import           Tests.Internal                     (testFunds, testFundsAll)
+import           Utils.ChainIndex                   (filterCleanUtxos)
 
 testFundsTS :: IO ()
 testFundsTS = testFunds @TestingServer
